CO-2947 move everything into cordapp module
