+++ .idea/compiler.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="CompilerConfiguration">
+    <bytecodeTargetLevel>
+      <module name="accounts.cordapp-client.integrationTest" target="1.8" />
+      <module name="accounts.cordapp-client.main" target="1.8" />
+      <module name="accounts.cordapp-client.test" target="1.8" />
+    </bytecodeTargetLevel>
+  </component>
+</project>
