+++ .idea/runConfiguration/WorldCup.xml
@@ -1,0 +1,16 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="World Cup" type="JetRunConfigurationType" factoryName="Kotlin">
+    <module name="accounts-trading-demo.agent" />
+    <option name="VM_PARAMETERS" value="" />
+    <option name="PROGRAM_PARAMETERS" value="--server.port=8081 --config.rpc.username=user1 --config.rpc.password=test --config.rpc.host=localhost --config.rpc.port=10005" />
+    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
+    <option name="ALTERNATIVE_JRE_PATH" />
+    <option name="PASS_PARENT_ENVS" value="true" />
+    <option name="MAIN_CLASS_NAME" value="net.corda.accounts.cordapp.sweepstake.clients.webserver.WorldCupApplication" />
+    <option name="WORKING_DIRECTORY" value="" />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
+</component>
+
