--- cordapp-client/src/main/kotlin/net/corda/accounts/cordapp/client/jsonparser/ResponseParser.kt
@@ -1,64 +1,0 @@
-package net.corda.accounts.cordapp.client.jsonparser
-
-import com.beust.klaxon.Klaxon
-import com.beust.klaxon.PathMatcher
-import org.slf4j.LoggerFactory
-import java.io.StringReader
-import java.util.regex.Pattern
-
-/**
- * Utility class for extracting data from the JSON response provided by a GitHub webhook.
- */
-class ResponseParser {
-
-    companion object {
-
-        private val logger = LoggerFactory.getLogger(ResponseParser::class.java)
-
-        /**
-         * Extracts the GitHub username from a JSON response when provided with the path the username should be
-         * provided from.
-         */
-        fun extractGitHubUsername(userNamePath: String, msg: String): String? {
-            var username: String? = null
-            try {
-                val pathMatcher = object : PathMatcher {
-                    override fun pathMatches(path: String) = Pattern.matches(userNamePath, path)
-                    override fun onMatch(path: String, value: Any) {
-                        if (value.toString().isNullOrBlank()) {
-                            throw IllegalArgumentException()
-                        } else {
-                            username = value.toString()
-                        }
-                    }
-                }
-                Klaxon().pathMatcher(pathMatcher).parseJsonObject(StringReader(msg))
-            } catch (e: IllegalArgumentException) {
-                logger.error("The request body does not contain a username at the specified path: $userNamePath.")
-                return null
-            }
-            return username
-        }
-
-        /**
-         * Retrieves the body of a comment and checks that it matches the 'createKey' phrase required to initiate the
-         * the [com.gitcoins.flows.CreateKeyFlow].
-         */
-        fun verifyCreateKey(msg: String): Boolean {
-            try {
-                val pathMatcher = object : PathMatcher {
-                    override fun pathMatches(path: String) = Pattern.matches(".*comment.*body.*", path)
-                    override fun onMatch(path: String, value: Any) {
-                        if (value.toString() != "createKey")
-                            throw IllegalArgumentException()
-                    }
-                }
-                Klaxon().pathMatcher(pathMatcher).parseJsonObject(StringReader(msg))
-            } catch (e: IllegalArgumentException) {
-                logger.error("Invalid pr comment. Please comment 'createKey'.")
-                return false
-            }
-            return true
-        }
-    }
-}
