--- cordapp-client/src/test/kotlin/net/corda/accounts/cordapp/client/webserver/SweepStakeControllerTests.kt
@@ -1,72 +1,0 @@
-package net.corda.accounts.cordapp.client.webserver
-
-import org.junit.Test
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNotNull
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.junit.runner.RunWith
-import org.springframework.boot.test.web.client.TestRestTemplate
-import org.springframework.http.HttpStatus
-import org.springframework.http.*
-import org.springframework.test.context.junit4.SpringRunner
-
-@RunWith(SpringRunner::class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class SweepStakeControllerTests {
-
-    @Autowired
-    lateinit var testRestTemplate: TestRestTemplate
-
-    @Test
-    fun `No endpoint`() {
-        val result = testRestTemplate.postForEntity("/blah", HttpEntity<String>(""), String::class.java)
-        assertEquals(HttpStatus.NOT_FOUND, result.statusCode)
-    }
-
-    @Test
-    fun `PR comment invalid`() {
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        val entity = HttpEntity<String>("""{ "comment": { "user": { "login": "testUser" }, "body": "Do nothing" } }""")
-        val result = testRestTemplate.postForEntity("/api/git/create-key", entity, String::class.java)
-        assertNotNull(result)
-        assertEquals(HttpStatus.BAD_REQUEST, result.statusCode)
-        assertEquals("Invalid pr comment. Please comment 'createKey'.", result.body.trim())
-    }
-
-    @Test
-    fun `PR comment createKey`() {
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        val entity = HttpEntity<String>("""{ "comment": { "user": { "login": "testUser1" }, "body": "createKey" } }""")
-        val result = testRestTemplate.postForEntity("/api/git/create-key", entity, String::class.java)
-        assertNotNull(result)
-        assertEquals(HttpStatus.CREATED, result.statusCode)
-        assertEquals("New public key generated for github user: testUser1", result.body.trim())
-    }
-
-    @Test
-    fun `Push event end point`() {
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        val entity = HttpEntity<String>("""{"pusher": { "name": "testUser1" } }""", headers)
-
-        val result = testRestTemplate.postForEntity("/api/git/push-event", entity, String::class.java)
-        assertNotNull(result)
-        assertEquals(HttpStatus.CREATED, result.statusCode)
-        assertEquals("New push event on the repo by: testUser1", result.body)
-    }
-
-    @Test
-    fun `Pull request event end point`() {
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        val entity = HttpEntity<String>("""{ "review": { "user": { "login": "testUser1" } } }""", headers)
-
-        val result = testRestTemplate.postForEntity("/api/git/pr-event", entity, String::class.java)
-        assertNotNull(result)
-        assertEquals(HttpStatus.CREATED, result.statusCode)
-        assertEquals("New pull request event on the repo by: testUser1", result.body.trim())
-    }
-}
