+++ cordapp-sweepstake/build.gradle
@@ -1,8 +1,11 @@
 apply plugin: 'kotlin'
 apply plugin: 'net.corda.plugins.cordapp'
 apply plugin: 'net.corda.plugins.cordformation'
-apply plugin: 'net.corda.plugins.quasar-utils'
-
+apply plugin: 'kotlin-jpa'
+apply plugin: 'java'
+apply plugin: 'idea'
+apply plugin: 'org.springframework.boot'
+apply plugin: 'io.spring.dependency-management'
 
 cordapp {
     info {
@@ -36,15 +39,30 @@ dependencies {
     cordaRuntime "$corda_release_group:corda:$corda_release_version"
     cordaCompile "$corda_release_group:corda-core:$corda_release_version"
 
+
     cordapp "$tokens_sdk:contract:$tokens_sdk_version"
     cordapp "$tokens_sdk:workflow:$tokens_sdk_version"
     cordapp "$tokens_sdk:money:$tokens_sdk_version"
 
+    compile "$corda_release_group:corda-rpc:$corda_release_version"
+    compile "$corda_release_group:corda-jackson:$corda_release_version"
+
+    compile("org.springframework.boot:spring-boot-starter-websocket:$spring_boot_version") {
+        exclude group: "org.springframework.boot", module: "spring-boot-starter-logging"
+    }
+
+    compile "org.apache.logging.log4j:log4j-slf4j-impl:${log4j_version}"
+    compile "org.apache.logging.log4j:log4j-web:${log4j_version}"
+    compile "org.slf4j:jul-to-slf4j:$slf4j_version"
+
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
 
     testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-    testCompile "junit:junit: 4.12"
+    testCompile "junit:junit:$junit_version"
     testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
+    testCompile (group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.0.2.RELEASE') {
+        exclude group: "org.springframework.boot", module: "spring-boot-starter-logging"
+    }
 }
 
 
@@ -60,8 +78,10 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
 
     nodeDefaults {
+        projectCordapp {
+            deploy = false
+        }
         cordapp project(":flows")
-//        cordapp project(":cordapp-client")
         cordapp project(":states-and-contracts")
     }
 
@@ -77,7 +97,7 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
         }
     }
     node {
-        name "O=Referee,L=London,C=GB"
+        name "O=PrizeManager,L=London,C=GB"
         p2pPort 10004
         sshdPort 2223
         rpcSettings {
@@ -86,16 +106,16 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
         }
         rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
     }
-    node {
-        name "O=PrizeManager,L=New York,C=US"
-        p2pPort 10008
-        sshdPort 2224
-        rpcSettings {
-            address("localhost:10009")
-            adminAddress("localhost:10010")
-        }
-        rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
-    }
+//    node {
+//        name "O=PrizeManager,L=New York,C=US"
+//        p2pPort 10008
+//        sshdPort 2224
+//        rpcSettings {
+//            address("localhost:10009")
+//            adminAddress("localhost:10010")
+//        }
+//        rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
+//    }
 //    node {
 //        name "O=FundManager2,L=Paris,C=FR"
 //        p2pPort 10012
@@ -121,4 +141,13 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
 task integrationTest(type: Test, dependsOn: []) {
     testClassesDirs = sourceSets.integrationTest.output.classesDirs
     classpath = sourceSets.integrationTest.runtimeClasspath
+}
+
+springBoot {
+    mainClassName = "net.corda.accounts.cordapp.sweepstake.clients.WorldCupApplicationKt"
+}
+
+task runPartyAServer(type: JavaExec, dependsOn: jar) {
+    classpath = sourceSets.main.runtimeClasspath
+    main = "net.corda.accounts.cordapp.sweepstake.clients.WorldCupApplicationKt"
 }
