+++ cordapp-sweepstake/src/main/kotlin/net/corda/accounts/cordapp/sweepstake/clients/SweepStakeController.kt
@@ -1,16 +1,18 @@
-package net.corda.accounts.cordapp.client.webserver
+package net.corda.accounts.cordapp.sweepstake.clients
 
+import net.corda.accounts.cordapp.sweepstake.clientsh.NodeRPCConnection
 import net.corda.accounts.cordapp.sweepstake.flows.Participant
 import net.corda.accounts.cordapp.sweepstake.flows.generateParticipantsFromFile
 import net.corda.accounts.cordapp.sweepstake.flows.generateTeamsFromFile
 import net.corda.core.utilities.contextLogger
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestMethod
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/api/sweepstake/")
-class SweepStakeController(rpc: NodeRPCConnection) {
+class SweepStakeController(@Autowired private val rpc: NodeRPCConnection) {
 
     companion object {
         private val logger = contextLogger()
@@ -22,21 +24,6 @@ class SweepStakeController(rpc: NodeRPCConnection) {
     fun createUser(): List<Participant> {
         return generateParticipantsFromFile("src/test/resources/participants.txt")
     }
-
-//    @RequestMapping("/create-accounts/", method = [RequestMethod.GET])
-//    fun createAccounts(): List<Participant> {
-//        // Create all the accounts for the players on node A
-//        val iterablePlayers = players.listIterator()
-//        while (iterablePlayers.hasNext()) {
-//            val player = iterablePlayers.next()
-//            if (!player.hasAccount) {
-//                proxyA.startFlow(net.corda.accounts.cordapp.sweepstake.flows::CreateAccountForPlayer, player).returnValue.getOrThrow()
-//                iterablePlayers.set(player.copy(hasAccount = true))
-//            }
-//        }
-//
-//        return generateParticipantsFromFile("src/test/resources/participants.txt")
-//    }
 
     @RequestMapping("/issue-teams/", method = [RequestMethod.GET])
     fun issueTeams(): List<Participant> {
