+++ cordapp-sweepstake/src/main/resources/application.properties
