+++ cordapp-sweepstake/src/main/resources/participants.txt
