+++ cordapp-sweepstake/src/main/resources/static/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Trustee</title>
+    <title>World Cup Sweepstake</title>
 </head>
 <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.11/lodash.min.js"></script>
 <style>
