+++ cordapp-sweepstake/src/main/resources/worldcupteams.txt
